JGC/mainWindowUpdates: Updated/arranged the current file structure to a bit more formal.
