+++ Trivial_Purfuit/Trivial_Pursuit.pyproject
@@ -1,11 +1,14 @@
 {
-    "files": ["mainwindow.py",
-              "ui_mainwindow.py",
-              "ui_start_menu.py",
-              "start_menu_wrapper.py",
-              "mainwindow.ui",
-              "start_menu.ui",
-              "player_token.py",
-              "the_board.py"
-              ]
+    "files": ["src/MainWindow/mainwindow.py",
+              "src/MainWindow/ui_mainwindow.py",
+              "src/MainWindow/mainwindow.ui",
+
+              "src/MainWindow/menus/ui_start_menu.py",
+              "src/MainWindow/menus/start_menu.ui",
+
+              "src/player_token/player_token.py",
+
+              "src/board/the_board.py",
+              "src/board/board_funcs.py"
+             ]
 }
