+++ Trivial_Purfuit/src/MainWindow/mainwindow.ui
