+++ Trivial_Purfuit/src/MainWindow/ui_mainwindow.py
